requirements.txt
flask==2.3.2
azure-storage-blob==12.21.1
pandas==2.0.3
openpyxl==3.1.2
gunicorn==20.1.0
